RayCastMesh Components; LargeVAO (Int); IntHolder > Int
The Meshes had more vertices than a short could hold. The indices are now stored in an Integer Array for meshes.
For more dynamical filling of meshes, IntHolder are now used for all addresses in the mesh generation.
Generic code has been written for solid and connector elements of a model.
The raycast mesh is done.