Adding Hotbar
Adding NanoVG framework for 2D drawing.
Adding Hotbar which draws the icons of the components.
Moving a bunch of interaction stuff for component selection from 3D to Hotbar (and route it).
Adjusting TextureWrapper to support textures with alpha.
Adding shared data class for sharing data between 2D and 3D planes.
Removed UglyComponentAwareness, which was in charge of simulating the TUNG component chooser.
New shader for rendering component icons in GUI.