Fix: Do cluster changes before simulating tick
I am not quite sure, why it was performed in the middle of the simulation, but in there it breaks clusters which may had been updated before editing.
Now it performs all edits before. If a component was sheduled for update, it now will fail in a way. But at least the structure makes more sense. If this causes issues they should be fixed in another way.