Store grab fields in one object; Fix parent removal on container grab
It is thread-safety unwise to store all the grab fields outside of one object. Also it makes it difficult to add more fields in that regards for container grabbing. Cleaner code on top.
The parent was removed, even if aborting to grab a container. No longer. Check moved in front of removal.