Rotate boards using quaternion and fix placement position
For more rotation angles (Y+XZ), the rotation is now stored as a quaternion, which gets modified by the rotation keys. Since this rotation is relative between faces, it does not need to change the over-the-edge-rotation-fix-code.
And offsets the position of the to be placed board along the placement normal vector, to position the board in a way that the edge of the grabbed board touches the parent board's face.