Refactor 3D-render class to prepare for improved placement
Fine-Placement sometimes relies on the rotation to be known. But the rotation was calculated multiple times at different places, but not at the time the placement position is getting calculated.
Thus this overdue refactoring is done. Its major goal is to calculate the placement data and placement rotation upfront. Sometimes with other data along.
This commit also fixes several mount-placement issues/bugs.
After this commit it will be much easier to insert the fine-placement.

A big part of this refactor is the new "Hitpoint" structure, which only gets updated once and stores the selected object up to the rotation/position things get placed on that object and if its on a board, it stores even the normal and position on that board face.
Several multi-thread accesses have been improved (not fully, but significantly reduced). Before it asked to break, now it tries not to.

Possible issues:
- The important data for each state in the render-class are still split in few fields, thus not fully thread-safe. (Side effects should be unnoticeable though).
- Wire and board drawing data is "hacked" into the Hitpoint structure.
- Currently the abort feature of placement is disabled, has to be added later again.