Add temporary version to jar
Since OpenTUNG does not have major or minor releases but just rolling updates with higher or lower priority, there is no version-number etc.
But there must be a version or state identifier in the distributed .jar files.
So now that will be git-<commitHash>, which gets printed whenever OpenTUNG starts.
Thus its fairly easy accessible for everyone. For now a "pre-official-version", good enough.