Fix AWT vs OpenGL fight
When creating a file-chooser before initializing OpenGL, OpenGL would fail eventually. Especially when another file-chooser is opened for saving files.
But when opening the file-chooser after initializing OpenGL the issue does not happen.
This only affects X11 system, so (some) Linux users. Works fine on Windows.
Thus when choosing the file-chooser on a Unix system, the dialog will now be delayed until OpenGL has been initialized.
The code for this not very elegant, but it will become obsolete with a proper GUI in OpenGL anyway.