Prevent multiple update of updateables
Currently only applies for components without user-interaction.
Should eventually be applied to clusters as well.
This is a better solution than the previous simulation tick counter variable.
It has the advantage, that when copying part of the simulation network, one can now easily check if a component would be updated in the next tick, which can then be copied as well.