Prevent mouse wheel stealing when selecting board
When the board was selected in the hotbar, the CONTROL key enabled fine placement and prevented mouse wheel control of the selected hotbar element.
This fix reduces the moments when this stealing happens to the cases where it actually can be applied.
It would also be possible to check if CONTROL is currently used on the movement keys. For now not checked.