Improve board drawing a lot
Board drawing used the center plane of the board (y=0), which means that the cursor was not on the square currently being dragged. Now the closest side of the board is being used.
Also when placing the board with a simple mouse click, the board growing code would trigger too early, and the user gets non 1×1 boards. With a simple mouse-click.
Now the growing is disabled, until the user moved the camera in any way.
The Hitpoint is now aware of the distance from camera to hitpoint.