Merge grabbing code for container and components
Due to growth reasons, these had been separated. The new copy code had them already merged.
However component grabbing needed a feature which container grabbing had.
So it only made sense to merge the two codes just like in copy.

As a side effect the grabbing of component with a wire to itself has been fixed (was a NPE). But copy placement is fully broken.
This is a WIP commit, it will break things.

The next commit will fix copy and grab, since the wires are collected on the wrong thread.