Increase resize bound rect a bit to allow less precision
Mounts are positioned directly on the edge of two squares, however if it is the wrong side of the edge, then the user cannot resize the board to directly under it, and sees 1 square too much.
Just increasing the bound rect a bit, will prevent the need to expand the board that extra square, since it is still within the bounds.