Change the way pre-saving halts the simulation and empties queues
Previously the pre-saving just stopped the ray-casting and disabled all modes and then just waited a very short duration for the render-queue to be empty. For simulation there was an atomic-integer which was checked with busy-waiting.

Now it does the same for stopping new tasks on the queues.
But the simulation now gets "locked" a more powerful pausing. And the queue emptying is confirmed by itself-injecting jobs, which only knowledge that the queue is empty after like 4-6 cycles of it being the only task in the queues. Still not perfect, but much better than before already. Due to this, busy-waiting has been removed.