Move all primary editing code into Tool classes
Every editing function was hacked into the main world render class.
Now each of them has been moved into its own class.

Only one primary tool can be on at a time.
If none is active, it uses input events mouse down and key up to detect if a tool has to go active.
If a tool is active, it forwards several input events into it.

A tool has to declare itself as "ready" which grants render and input thread access to it.
A tool has to turn input and render thread access off separately.