Add multi-wire-placing
The most basic form of multi-wire-placing has been added.
Along this change, tools now receive the right-click-down and the hitpoint does not longer get abused as temporary wire storage.
Ray-Cast has new code which is just in charge of collecting connectors. Could have been more generic. But with the next and different use-cases that is a better opportunity.