Fix wire connection point for Displays
Displays and Noisemaker and maybe more components have a non-default position for the wire-connection point.
By default that is 0.9 far away from the component.
But the displays are standing on the pegs and thus it must be closer to the component's center.
However it is a custom value.
That value can now be defined per peg in the model.
However the value is currently stored in the Meshable, where it does not belong.