Fix capitalization in key manager system and related
Either the readable value was not processed at all. Or incorrectly in some cases.
Either way, now there is one function to be called whenever a new readable value was acquired. And it is applied in all cases.
Means single-letters are now always capital letters.
And on Windows, the empty string gets converted to null. Which improves the output in the GUI.