Highlight connectors when MWP between groups
After you placed the first group of connectors, and before you started selecting the second group.
Your mouse was not down and freely looking around also at connectors but without highlighting.

Now only the connectors get highlighted in that time. That way it should be more easy to detect one is in the wire-drawing mode and where to perform the next action.