Enhance cluster highlighter
Fixes that deleting connectors does not mark the connected wire as deleted (as in their parents are not set to null).

The cluster highlighter shows connected clusters now.
If these are drain-clusters it shows them fully.
If source-clusters then only the wire to the blot and the blot itself.
It adds the ability for the cluster-highlighter to regenerate its to highlighted content.
Hides the clusters while grabbing a container though. Since that could take too much performance.