Add ability to edit labels using a Swing window
Additional do a bit of refactoring to label texture management.
Now it may only be modified from the render thread.
Only one method to change the texture of a label. And 'null' usage textures can never be unloaded while runtime.
Textures won't be applied, if the label is deleted.

Editing tool, now opens the LabelEditor to edit label text.