Fix rounding vector normalization
Sometimes the vector appears to have length 0, then it cannot be normalized, cause that would mean division by 0. In these cases we skip the rounding.
Happens, in general this goes along with a 0° angle.
Additional NaN will from now on be detected and the grabment will be aborted.
Since if it happens it causes several unpredictable side effects, which take time to trace. (Wire raycaster looping until GC-Overhead).