Fix 'hasTexture()' returning always true
'hasTexture()' checked if the text variable was null.
By convention that will never be null. It is an empty string.
To prevent confusion 'hasText()' was added.
'hasTexture()' is now used whenever a label is about to be drawn, this prevents the label to have the texture of the last label.