Add import to grabbing tool
The import tool has been added. As the label-editor it uses JavaSwing (unsafe) until interfaces have been improved.
The import tool has its own window but is in charge of the import. It loads the board using the loader class and then post-processes the board, just like BoardUniverse and copyImpl() in grabbing tool.
As last step it replaces the active tool and hands over the control to the grabbing tool to finalize the import by doing the rest of copyImpl().